Resume_Shortlister/requirements.txt
streamlit 
google-generativeai
PyPDF2
python-dotenv
stqdm